Day 1 complete.

+++ src/day.rs
@@ -1,4 +1,4 @@
 pub trait Day {
-    fn part1(&self) -> Result<i64, &str> ;
-    fn part2(&self) -> Result<i64, &str> ;
+    fn part1(&self) -> Result<usize, &str> ;
+    fn part2(&self) -> Result<usize, &str> ;
 }
